Fixes CRM-10120.
Permit API key authentication when $params['name'] is set.
